fix ascendc_parse_dumpinfo.py softlink

+++ scripts/package/scripts/asc-devkit_custom_create_softlink.sh
@@ -71,7 +71,11 @@ do_create_stub_softlink() {
         if [ ! -d "$install_path/$latest_dir/tools/ascendc_tools" ]; then
             mkdir -p "$install_path/$latest_dir/tools/ascendc_tools"
         fi
-        ln -sr "$install_path/$version_dir/tools/ascendc_tools/ascendc_parse_dumpinfo.py" "$install_path/$latest_dir/tools/ascendc_tools/ascendc_parse_dumpinfo.py"
+        if [ ! -L "$install_path/$latest_dir/tools/ascendc_tools/ascendc_parse_dumpinfo.py" ]; then
+            if [ -e "$install_path/$version_dir/tools/ascendc_tools/ascendc_parse_dumpinfo.py" ]; then
+                ln -sr "$install_path/$version_dir/tools/ascendc_tools/ascendc_parse_dumpinfo.py" "$install_path/$latest_dir/tools/ascendc_tools/ascendc_parse_dumpinfo.py"
+            fi
+        fi
     fi
     if [ -d "$arch_linux_path/pkg_inc/asc/hccl" ]; then
         chmod 750 "$arch_linux_path/pkg_inc/asc/hccl"
